chore: refactored into ssr and csr components
