+++ app/components/NavBar.tsx
@@ -3,9 +3,11 @@ import Link from "next/link";
 const NavBar = () => {
   return (
     <nav className="bg-white p-2 flex justify-between">
-      <Link href="/" className="font-bold text-gray-700 text-2xl">
-        OpenTable
-      </Link>
+      <h1>
+        <Link href="/" className="font-bold text-gray-700 text-2xl">
+          OpenTable
+        </Link>
+      </h1>
       <div>
         <div className="flex">
           <button className="bg-blue-400 text-white border p-1 px-4 rounded mr-3">
