+++ app/restaurant/[slug]/components/RestaurantHeader.tsx
@@ -5,9 +5,9 @@ const RestaurantHeader = () => {
     <div className="h-96 overflow-hidden">
       <NavBar />
       <div className="bg-center bg-gradient-to-r from-[#0f1f47] to-[#5f6984] h-full flex justify-center items-center">
-        <h1 className="text-7xl text-white captitalize text-shadow text-center">
+        <h2 className="text-7xl text-white captitalize text-shadow text-center">
           Milestones Grill (Toronto)
-        </h1>
+        </h2>
       </div>
     </div>
   );
