+++ app/restaurant/[slug]/components/RestaurantTitle.tsx
@@ -1,7 +1,7 @@
 const RestaurantTitle = () => {
   return (
     <div className="mt-4 border-b pb-6">
-      <h2 className="font-bold text-6xl">Milesstone Grill</h2>
+      <h3 className="font-bold text-6xl">Milesstone Grill</h3>
     </div>
   );
 };
